handling input using directInput and observer pattern is done

+++ Module 00/Assignment00_DirectX/Assignment00_DirectX/InputObserver.hpp
@@ -3,7 +3,7 @@
 class InputObserver
 {
 	public:
-		virtual void	keyDown(int key) = 0;
+		virtual void	keyDown(BYTE key) = 0;
 		virtual void	keyUp(int key) = 0;
 		virtual	void	mouseMove(float x, float y) = 0;
 };
